+++ tests/integration/taskrouter/v1/workspace/worker/test_worker_channel.py
@@ -37,9 +37,9 @@ class WorkerChannelTestCase(IntegrationTestCase):
                     "key": "channels",
                     "next_page_url": null,
                     "page": 0,
-                    "page_size": 1,
+                    "page_size": 50,
                     "previous_page_url": null,
-                    "url": "https://taskrouter.twilio.com/v1/Workspaces/WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Workers/WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Channels"
+                    "url": "https://taskrouter.twilio.com/v1/Workspaces/WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Workers/WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Channels?PageSize=50&Page=0"
                 },
                 "channels": [
                     {
@@ -48,12 +48,12 @@ class WorkerChannelTestCase(IntegrationTestCase):
                         "available": true,
                         "available_capacity_percentage": 100,
                         "configured_capacity": 1,
-                        "date_created": "2014-05-14T10:50:02Z",
-                        "date_updated": "2014-05-15T16:03:42Z",
+                        "date_created": "2016-04-14T17:35:54Z",
+                        "date_updated": "2016-04-14T17:35:54Z",
                         "sid": "WCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                         "task_channel_sid": "TCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                         "task_channel_unique_name": "default",
-                        "url": "https://taskrouter.twilio.com/v1/Workspaces/WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Workers/WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Channels/WRaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                        "url": "https://taskrouter.twilio.com/v1/Workspaces/WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Workers/WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Channels/WCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                         "worker_sid": "WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                         "workspace_sid": "WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
                     }
@@ -78,9 +78,9 @@ class WorkerChannelTestCase(IntegrationTestCase):
                     "key": "channels",
                     "next_page_url": null,
                     "page": 0,
-                    "page_size": 1,
+                    "page_size": 50,
                     "previous_page_url": null,
-                    "url": "https://taskrouter.twilio.com/v1/Workspaces/WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Workers/WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Channels"
+                    "url": "https://taskrouter.twilio.com/v1/Workspaces/WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Workers/WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Channels?PageSize=50&Page=0"
                 },
                 "channels": []
             }
@@ -106,7 +106,7 @@ class WorkerChannelTestCase(IntegrationTestCase):
             'https://taskrouter.twilio.com/v1/Workspaces/WSXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX/Workers/WKXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX/Channels/WCXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX',
         ))
 
-    def test_fetch_response(self):
+    def test_fetch_sid_response(self):
         self.holodeck.mock(Response(
             200,
             '''
@@ -116,12 +116,12 @@ class WorkerChannelTestCase(IntegrationTestCase):
                 "available": true,
                 "available_capacity_percentage": 100,
                 "configured_capacity": 1,
-                "date_created": "2014-05-14T10:50:02Z",
-                "date_updated": "2014-05-15T16:03:42Z",
+                "date_created": "2016-04-14T17:35:54Z",
+                "date_updated": "2016-04-14T17:35:54Z",
                 "sid": "WCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "task_channel_sid": "TCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "task_channel_unique_name": "default",
-                "url": "https://taskrouter.twilio.com/v1/Workspaces/WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Workers/WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Channels/WRaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "url": "https://taskrouter.twilio.com/v1/Workspaces/WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Workers/WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Channels/WCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "worker_sid": "WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "workspace_sid": "WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
             }
@@ -157,8 +157,8 @@ class WorkerChannelTestCase(IntegrationTestCase):
                 "available": true,
                 "available_capacity_percentage": 100,
                 "configured_capacity": 3,
-                "date_created": "2014-05-14T10:50:02Z",
-                "date_updated": "2014-05-15T16:03:42Z",
+                "date_created": "2016-04-14T17:35:54Z",
+                "date_updated": "2016-04-14T17:35:54Z",
                 "sid": "WCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "task_channel_sid": "TCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "task_channel_unique_name": "default",